refactor: Squeezing related solvers in one file respectively
