--- biotrainer/solvers/residues_classification_solver.py
@@ -1,6 +1,0 @@
-from .sequence_classification_solver import SequenceClassificationSolver
-
-
-# Identical to the SequenceClassificationSolver at the moment
-class ResiduesClassificationSolver(SequenceClassificationSolver):
-    pass
