--- biotrainer/solvers/residues_regression_solver.py
@@ -1,6 +1,0 @@
-from .sequence_regression_solver import SequenceRegressionSolver
-
-
-# Identical to the SequenceRegressionSolver at the moment
-class ResiduesRegressionSolver(SequenceRegressionSolver):
-    pass
