+++ biotrainer/solvers/residues_solvers.py
@@ -1,0 +1,10 @@
+from .sequence_solvers import SequenceClassificationSolver, SequenceRegressionSolver
+
+
+# Identical to the SequenceClassificationSolver at the moment
+class ResiduesClassificationSolver(SequenceClassificationSolver):
+    pass
+
+# Identical to the SequenceRegressionSolver at the moment
+class ResiduesRegressionSolver(SequenceRegressionSolver):
+    pass
