--- biotrainer/solvers/sequence_regression_solver.py
@@ -1,8 +1,0 @@
-import torch
-
-from .solver import Solver
-
-
-class SequenceRegressionSolver(Solver):
-    def _transform_network_output(self, network_output: torch.Tensor) -> torch.Tensor:
-        return network_output.flatten().float()
